backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
openai==1.59.7
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
cryptography==44.0.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
certifi==2024.12.14
supabase==2.11.0

# Resume parsing
PyMuPDF==1.25.1
pdfplumber==0.11.5
python-docx==1.1.2
reportlab==4.2.5
qrcode[pil]==8.0

ruff==0.8.6

# Ops: rate limiting, response caching helpers, optional error monitoring
slowapi==0.1.9
cachetools==5.5.0
sentry-sdk[fastapi]==2.19.2
